Make de-slanting, spectrum extraction, and calibration more robust.
For marginal (as in badly focused) images the de-slanting runs into an
exception because it can't find any peaks: limp on by not doing
slant correction.

Likewise the spectrum extraction can't separate the signal from the
sky with fixed settings: try with iteratively increasing limits until
some sort of sky/signal separation is found.

In calibration spectra peaks tend to be very wide: try with varying
peak widths to find a minimum number of peaks.